% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_workdays.R
\name{count_workdays}
\alias{count_workdays}
\title{Calculate total pay between two dates}
\usage{
count_workdays(first, last)
}
\arguments{
\item{first}{Atomic string. First pay date (yyyy-mm-dd).}

\item{last}{Atomic string. Last pay date (yyyy-mm-dd).}

\item{daily}{Atomic double.}
}
\value{
Atomic double.
}
\description{
Email on Wed, May 26, 2021 at 8:12 PM: "The daily rates that were calculated
from our annualized salary excluded vacation / holiday days.  To get around a
recalculation, \strong{we will submit monthly invoices in which the amount due will
be the number of weekdays within the invoice period, irrespective of holidays
or vacation days spent}. This means that our first invoice will cover the
time period of April 28th to May 27th and our monthly rate for this time
period will be 22 days (22 * daily rate as documented in our contracts)."
}
\examples{
first <- "2021-05-28"
last <- "2021-06-27"
count_workdays(first = "2021-05-28", last = "2021-06-27")
}
