% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_workdays.R
\name{count_workdays}
\alias{count_workdays}
\title{Calculate total pay between two dates}
\usage{
count_workdays(first, last)
}
\arguments{
\item{first}{Atomic string. First pay date (yyyy-mm-dd).}

\item{last}{Atomic string. Last pay date (yyyy-mm-dd).}

\item{daily}{Atomic double.}
}
\value{
Atomic double.
}
\description{
Calculate total pay between two dates
}
\examples{
first <- "2021-05-28"
last <- "2021-06-27"
count_workdays(first = "2021-05-28", last = "2021-06-27")
}
